all weights are of type double now

+++ graph/include/graph.h
@@ -13,9 +13,8 @@ typedef struct _graph graph;
 // объявление переменной, содержащей значение,
 // сигн. об отсутствии ребра в структуре данных
 // определена в файлах исходного кода структур данных
-// как 0 для невзвешенных графов
-// и как INT_MAX для взвешенных графов
-extern const int no_edge;
+// как INFINITY
+extern const double no_edge;
 
 graph *make_graph(size_t, unsigned char);
 graph *copy_graph(graph *);
@@ -24,11 +23,11 @@ void dest_graph(graph *);
 size_t vertices(graph *);
 unsigned char directed(graph *);
 
-int add_edge(graph *, size_t, size_t, int);
-int get_edge(graph *, size_t, size_t);
+int add_edge(graph *, size_t, size_t, double);
+double get_edge(graph *, size_t, size_t);
 int remove_edge(graph *, size_t, size_t);
 
 void transpose(graph *);
-int *make_adj_mat(graph *);
+double *make_adj_mat(graph *);
 
 #endif
